mantle: implement VK_KHR_dynamic_rendering
the current renderpass implementation has a few issues:
- image layouts are not known at pipeline creation, so we can't use
  optimal layouts in the renderpass
- targets declared at pipeline creation sometimes differ from the
  targets bound at draw time (illegal, but seen in BF4)
- we can't use OP_CLEAR to clear targets at the beginning of the
  renderpass.

solving these would require deferred renderpass creation and a cache.
however this extension removes the need for renderpasses entirely, which
simplifies the code quite a bit.

older GCN GPUs won't be compatible as they are now unsupported on Windows,
but the official Mantle driver can be used in this case.

register target formats at draw time to solve the 2nd issue.