mantle: remove GPU->CPU->GPU sync on initial image transition
BF4 was very stuttery when moving around, because the game creates
non-target images on the fly, and we had a huge pipeline bubble when
transitioning these images to the initial data transfer state.

we were waiting for the barriers to complete before submitting
further work but it's not necessary because the barriers will block
execution as needed. (that's what they're for...)

use a pool of command buffers so we don't have to wait for the previous
one to be cmd buffer to be available.