mantle: always assume 8 color attachments
this makes things simpler and reduces the risk for color format or blend
state mismatch, which were previously assumed to be game bugs.

attachments can be declared as unused by passing VK_NULL_HANDLE in place
of the image view in vkBeginRenderingKHR, or by passing an undefined color
format in VkPipelineRenderingCreateInfoKHR at pipeline creation.